app/backendserver/College_Rag/requirements.txt
langchain
langchain-community
langchain-core
langchain-text-splitters
langchainhub
chromadb
sentence-transformers
python-dotenv
